+++ sources/RI.Abstractions.Tests/Composition/ICompositionContainer_Register.cs
@@ -16,6 +16,7 @@ namespace RI.Abstractions.Tests.Composition
         [MemberData(nameof(ICompositionContainer_Register.GetContainers))]
         public void Register_Null_ArgumentNullException(ICompositionContainer instance)
         {
+            // Act + Assert
             Assert.Throws<ArgumentNullException>(() => instance.Register(null));
         }
 
@@ -23,6 +24,7 @@ namespace RI.Abstractions.Tests.Composition
         [MemberData(nameof(ICompositionContainer_Register.GetContainers))]
         public void Register_Empty_NoException(ICompositionContainer instance)
         {
+            // Act
             instance.Register(new CompositionRegistration[0]);
         }
 
@@ -30,7 +32,9 @@ namespace RI.Abstractions.Tests.Composition
         [MemberData(nameof(ICompositionContainer_Register.GetContainersAndRegistrations))]
         public void Register_ValidRegistrations_NoException(ICompositionContainer instance, CompositionRegistration[] validRegistrations)
         {
-            Assert.True(!instance.Registered);
+
+            // Act + Assert
+            Assert.False(instance.Registered);
             instance.Register(validRegistrations);
             Assert.True(instance.Registered);
         }
@@ -39,14 +43,15 @@ namespace RI.Abstractions.Tests.Composition
         [MemberData(nameof(ICompositionContainer_Register.GetContainers))]
         public void Register_Reregister_InvalidOperationException(ICompositionContainer instance)
         {
+            // Act + Assert
             instance.Register(new CompositionRegistration[0]);
             Assert.Throws<InvalidOperationException>(() => instance.Register(new CompositionRegistration[0]));
         }
 
         public static IEnumerable<object[]> GetContainers() =>
-            ICompositionContainerTestFactory.GetContainers();
+            _CompositionTestFactory.GetContainers();
 
         public static IEnumerable<object[]> GetContainersAndRegistrations() =>
-            ICompositionContainerTestFactory.GetContainersAndRegistrations();
+            _CompositionTestFactory.GetContainersAndRegistrations();
     }
 }